assets/scripts/requirements.txt
optigatrust
pyOpenSSL
py7zr